Move Payment class to domain directory
