+++ app/src/main/java/angulo/javier/myfinpal/ui/home/HomeFragment.kt
@@ -14,7 +14,7 @@ import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import angulo.javier.myfinpal.databinding.FragmentHomeBinding
-import angulo.javier.myfinpal.entity.Payment
+import angulo.javier.myfinpal.domain.Payment
 import angulo.javier.myfinpal.ui.adapter.PaymentAdapter
 import angulo.javier.myfinpal.util.Categories
 import angulo.javier.myfinpal.util.PaymentMethods
