+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -53,7 +53,7 @@
 
         <argument
             android:name="payment"
-            app:argType="angulo.javier.myfinpal.entity.Payment" />
+            app:argType="angulo.javier.myfinpal.domain.Payment" />
         <action
             android:id="@+id/action_navigation_history_detail_to_navigation_history"
             app:destination="@id/navigation_history"
